перенёс настройки карты в OffersMap
